requirements.txt
Pillow==7.0.0
playsound==1.2.2
pyobjc==6.2